Manually remove the older-style properties in the ImageAnalysisRequest object to simplify client generation.
The API specifies multiple sets of properties that can be used to achieve the same thing (for backwards compatibility), but
that causes issues with client generation in go as it doesn't expect some to be null. This change uses only one way of
specifying the image to analyze and simplifies the generation of the client.

No functionality is lost, just clarity in the API since the user is interacting with the generated code this ensures code will
only use the newer method and be more future proof.

Signed-off-by: Zach Hill <zach@anchore.com>